support-agent/requirements.txt
langchain>=0.3.0
langchain-groq>=0.1.0
langchain-community>=0.3.0
chromadb>=0.5.5
sentence-transformers>=3.0.1
streamlit>=1.38.0
python-dotenv>=1.0.1
pandas>=2.2.2
plotly>=5.24.1
